move collection sidebar into its own dir
