+++ src/components/GalleryEditor/CollectionSidebar/CollectionListItem.tsx
