+++ src/components/GalleryEditor/CollectionSidebar/CreateCollectionIcon.tsx
